requirements.txt
streamlit
torch
torchvision
torchaudio
ultralytics
opencv-python
Pillow
numpy
